wjx_put1

+++ src/main/java/com/hwadee/ssm/entity/WorkOvertime.java
@@ -30,7 +30,7 @@ public class WorkOvertime implements Serializable {
   }
 
 
-  public long getType() {
+  public int getType() {
     return type;
   }
